Add files via upload
This script checks for a device to be on the network. When it disconnects
the script minimizes currently opened windows, locks the workstation and 
opens Internet Explorer (I.E.) to play music or live stream. Once the 
device returns to the network I.E. shuts down and the windows are un-minimized.